+++ src/mkvtoolnix-gui/merge/source_file.h
@@ -71,11 +71,6 @@ public:
 
 protected:
   virtual mtx::bcp47::language_c deriveLanguageFromFileName();
-  virtual QRegularExpression regexForDerivingLanguageFromFileName();
-
-public:
-  static QString defaultRegexForDerivingLanguageFromFileName();
-  static void setupFromPreferences();
 };
 
 }