+++ src/mkvtoolnix-gui/merge/tool.cpp
@@ -128,7 +128,6 @@ Tool::setupActions() {
 
   connect(mw,                                         &MainWindow::preferencesChanged,   this, [&p]() { Util::setupTabWidgetHeaders(*p.ui->merges); });
   connect(mw,                                         &MainWindow::preferencesChanged,   this, &Tool::retranslateUi);
-  connect(mw,                                         &MainWindow::preferencesChanged,   this, []() { SourceFile::setupFromPreferences(); });
 
   connect(App::instance(),                            &App::addingFilesToMergeRequested, this, &Tool::identifyMultipleFilesFromCommandLine);
   connect(App::instance(),                            &App::openConfigFilesRequested,    this, &Tool::openMultipleConfigFilesFromCommandLine);